requirements.txt
# Core Flask server
flask==3.0.0
flask-cors==4.0.0

# Audio processing
pyaudio==0.2.14
pydub==0.25.1
soundfile==0.12.1

# AI/ML libraries
torch==2.1.0
torchaudio==2.1.0

# Speech recognition and TTS
faster-whisper==0.10.0
elevenlabs==0.2.27
openai-whisper==20231117

# Claude AI integration
boto3==1.34.0
langchain-aws==0.1.0
langchain-core==0.1.0
langchain==0.1.0

# Computer vision and screen capture
opencv-python==4.8.1.78
mss==9.0.1
pillow==10.1.0
numpy==1.24.3

# Windows integration (for Audio2Face window detection)
pywin32==306

# HTTP requests and utilities
requests==2.31.0
python-dotenv==1.0.0

# Data handling
pandas==2.1.4
emoji==2.8.0

# Development and debugging
python-dateutil==2.8.2
certifi==2023.11.17
charset-normalizer==3.3.2
idna==3.6
urllib3==2.1.0

# Optional: For enhanced audio processing
librosa==0.10.1
scipy==1.11.4

# Optional: For better error handling and logging
colorama==0.4.6
tqdm==4.66.1